introduce FocusTextField
